% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmp_client.R
\name{fmp_fx_quotes}
\alias{fmp_fx_quotes}
\title{Get all Forex quotes (v3 & stable)}
\usage{
fmp_fx_quotes(version = c("v3", "stable"))
}
\arguments{
\item{version}{"v3" oder "stable"}
}
\value{
Tibble mit mind. \code{symbol}, optional \code{price}, \code{bid}, \code{ask}, \code{time}, …
}
\description{
Holt alle verfügbaren FX-Paare mit aktuellen Quotes.
v3: /api/v3/quotes/forex
stable: /stable/batch-forex-quotes
}
