% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmpbrunch-package.R
\docType{package}
\name{fmpbrunch}
\alias{fmpbrunch}
\alias{fmpbrunch-package}
\title{fmpbrunch: Minimal FMP Client (v3 & stable)}
\description{
Light R-Client for FinancialModelingPrep. Support prices,
Metrics (key/ratios/growth/ev), profiles and more – everything with tidy tibbles.
}
\details{
\strong{Funktionen:}
\itemize{
\item \code{fmp_prices()}  – historical prices (v3 & stable)
\item \code{fmp_metrics()} – Metrics (key/ratios/growth/ev; TTM optional; v3 & stable)
\item \code{fmp_profile()} – company profiles (v3 & stable)
\item FX: \code{fmp_fx_quotes()} - All available currency pairs
\item Crypto: \code{fmp_crypto_quotes()} - All available cryptos
\item Commodities: \code{fmp_commodity_quotes()} - All avaialble commodities
\item Exchanges: \code{fmp_exchanges()} - All available exchanges
}

\strong{Quick Start}
\preformatted{
library(fmpbrunch)
set_fmp_key("YOUR_KEY")
fmp_prices(c("AAPL","MSFT"), from = "2025-10-15", version = "stable")
fmp_metrics("AAPL", metric = "key", ttm = TRUE, version = "stable")
fmp_profile("AAPL", version = "stable")
fmp_fx_quotes("v3")
fmp_crypto_quotes("v3")
fmp_commodity_quotes("stable")
fmp_exchanges("stable")
}
}
\seealso{
\code{\link[=fmp_prices]{fmp_prices()}}, \code{\link[=fmp_metrics]{fmp_metrics()}}, \code{\link[=fmp_profile]{fmp_profile()}}
}
\author{
\strong{Maintainer}:

Authors:
\itemize{
  \item Börsenbrunch \email{boersenbrunch@gmail.com}
}

}
\keyword{internal}
