% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmpbrunch-package.R
\docType{package}
\name{fmpbrunch}
\alias{fmpbrunch}
\alias{fmpbrunch-package}
\title{fmpbrunch: Minimal FMP Client (v3 & stable)}
\description{
Leichter R-Client für FinancialModelingPrep. Unterstützt Kurse,
Metrics (key/ratios/growth/ev) und Profile – alles als tidy tibbles.
}
\details{
\strong{Funktionen:}
\itemize{
\item \code{fmp_prices()}  – historische Preise (v3 & stable)
\item \code{fmp_metrics()} – Kennzahlen (key/ratios/growth/ev; TTM optional; v3 & stable)
\item \code{fmp_profile()} – Firmenprofile (v3 & stable)
\item FX: \code{fmp_fx_quotes()} - Alle verfügbaren Währungspaare und aktuelle Kurse
\item Crypto: \code{fmp_crypto_quotes()} - Alle verfügbaren Crypto Währungen und aktuelle Kurse
\item Commodities: \code{fmp_commodity_quotes()} - Alle verfügbaren Rohstoffe und aktuelle Kurese
\item Exchanges: \code{fmp_exchanges()} - Alle Verfügbaren Börsen
}

\strong{Quick Start}
\preformatted{
library(fmpbrunch)
set_fmp_key("YOUR_KEY")
fmp_prices(c("AAPL","MSFT"), from = "2025-10-15", version = "stable")
fmp_metrics("AAPL", metric = "key", ttm = TRUE, version = "stable")
fmp_profile("AAPL", version = "stable")
fmp_fx_quotes("v3")
fmp_crypto_quotes("v3")
fmp_commodity_quotes("stable")
fmp_exchanges("stable")
}
}
\seealso{
\code{\link[=fmp_prices]{fmp_prices()}}, \code{\link[=fmp_metrics]{fmp_metrics()}}, \code{\link[=fmp_profile]{fmp_profile()}}
}
\author{
\strong{Maintainer}: 

Authors:
\itemize{
  \item Martin Kiontke \email{boersenbrunch@gmail.com}
}

}
\keyword{internal}
